backend/traffic-routing-service/requirements.txt
fastapi
uvicorn
osmnx
networkx
kafka-python
scipy
geopy
pydantic
requests
Rtree
scikit-learn
